pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.1",
    "textblob>=0.19.0",
    "uvicorn>=0.35.0",
    "vadersentiment>=3.3.2",
    "yfinance>=0.2.65",
]
